Updated to newest Google Gemini 2.5 pro model
Updated gemini 2.5 pro to newest version number, think this should be done automatically, but why not

Also added gemini 2.5 flash, did this really quickly because i prefer this models when testing myself, should add a thinking version using thinkingbudet as well, but did this in 2 secs, for myself.

Have not seen how the provider map or types is supposed to work, but just throwed them in there.